requirements.txt
google-cloud-error-reporting~=1.11.1
fastapi~=0.115.8
fastapi-cli~=0.0.7
Flask>=2.2.2
numpy~=2.2.3
pandas~=2.2.3
pydantic~=2.10.6
xgboost~=2.1.4
scikit-learn~=1.6.1
uvicorn~=0.34.0
